pyproject.toml
[project]
name = "peters"
version = "0.0.1"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.16.3",
    "fastapi>=0.116.0",
    "fastapi-async-sqlalchemy>=0.6.1",
    "fastapi-storages>=0.3.0",
    "pydantic-settings>=2.10.1",
    "sqladmin>=0.21.0",
    "sqlmodel>=0.0.24",
    "uv>=0.7.19",
]

[tool.black]
line-length = 100 
target-version = [ "py310", "py311" ]
exclude = "((.eggs | .git | .pytest_cache | build | dist))"

[tool.ruff]
line-length = 100 
exclude = [".git", "__pycache__", ".mypy_cache", ".pytest_cache"]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = [
    "B904",
    "B006",    
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = ["alembic", "__pycache__"]
